api/requirements.txt
# Common
psycopg2-binary==2.8.6
django==3.2.5
djangorestframework==3.11.2
django-cors-headers==3.4.0
whitenoise==5.1.0
requests==2.24.0
Jinja2==2.11.3
markupsafe==2.0.1
pyyaml==5.3.1
uritemplate==3.0.1

# Datagrowth
jsonschema==3.2.0
pytz==2020.1
urlobject==2.4.3
pandas==1.1.5
git+https://github.com/fako/django-json-field.git
lxml==4.6.1
datagrowth==0.17.6

# Tools
invoke==1.4.1
sentry-sdk==1.0.0

# AWS
awscli==1.19.85
boto3==1.17.85
requests-aws4auth==1.0
